Xor before mixing.
It turns out the x86 AES instructions xor the key *after* mixing, which
isn't what we want.  That unfortunately means we need to explicitly xor
our data in prior to running the first AES round, which is a bit of a
performance hit.  Not too bad, but it is annoying.