Made a silly mistake in finalization that gave false performance.
Xoring some of the state into the final result was accidentally
skipped, and the compiler was smart enough to infer that some earlier
processing didn't need to be done because of that.  That's what I get
for not checking the assembly more closely, or just doing general sanity
checking.

In any case, it seems like 50 GB/s may not be achievable while
maintaining unambiguously high hash quality.  At least, not with this
simple construction.  Nevertheless, this is still achieving 40 GB/s,
which is faster than I anticipated after discovering the bug.  Maybe
there are still bugs lurking.  But I simplified the code by using
explicit loops again, to make it more obvious when things are weird, so
hopefully no more bugs. Fingers crossed.